Move nextgen fixture to cloudformation/conftest.py
